create Question function done
